build: Don't run ldconfig when installing into a DESTDIR
If we're installing into a DESTDIR, for example during build of a
package, then there is no point in re-running ldconfig, because the
configuration file /etc/ld.so.conf.d/libstrangle.conf is not visible
to ldconfig yet. Also, builds with a DESTDIR will typically be done as
non-root, so we will not have the necessary permissions to update the
ld.so.cache anyway.

This makes one of the Arch Linux packaging patches unnecessary, so
remove it.

Signed-off-by: Simon McVittie <smcv@collabora.com>