real_dlsym: Crash out with an error if we cannot find the real dlsym
Previously, libstrangle would crash immediately after this with a
null pointer dereference (segmentation fault). There's no real way to
recover in this situation, so failure is the only option, but printing
an error message on stderr and aborting with a core dump seems like a
more user-friendly way to fail.

Signed-off-by: Simon McVittie <smcv@collabora.com>